This should've been checked in with the header 'gmsIndexFile.h'
See that comment ... about 'long' is 64-bit and DCW is only
8, 16, or 32 bit.  Any 64 bit causes parsing to bomb.

Changes to be committed:
	modified:   source/gmsFile/gmsIndexFile.cpp

+++ C_source/source/gmsFile/gmsIndexFile.cpp
@@ -163,12 +163,12 @@ void gmsPrintIndexRecord
    for (i = 0;
         i < theIndexTable->numberRecordsInTable;
         i++)
-      {
-       printf("  %d \t %ld \t\t %d\n",
+   {
+       printf("  %d \t %d \t\t %d\n",
               i,
               theIndexTable->indexList[i].byteOffset,
               theIndexTable->indexList[i].numBytes);
-      }
+   }
 
    printf("______________________________\n");
 }
@@ -212,10 +212,9 @@ static void buildIndexRecord
        numChars);
 
    for (i = 0; i < theIndexTable->numberRecordsInTable; i++)
-      {
-       theIndexTable->indexList[i].byteOffset =
-          (long) gmsReadInteger (index_fd);
+   {
+       theIndexTable->indexList[i].byteOffset = gmsReadInteger (index_fd);
 
        theIndexTable->indexList[i].numBytes = gmsReadInteger (index_fd);
-      }
+   }
 }
